+++ src/routing/or.rs
@@ -47,11 +47,13 @@ where
     }
 
     fn call(&mut self, req: Request<ReqBody>) -> Self::Future {
+        let original_uri = req.uri().clone();
         ResponseFuture {
             state: State::FirstFuture {
                 f: self.first.clone().oneshot(req),
             },
             second: Some(self.second.clone()),
+            original_uri: Some(original_uri),
         }
     }
 }
@@ -66,6 +68,9 @@ pin_project! {
         #[pin]
         state: State<A, B, ReqBody>,
         second: Option<B>,
+        // Some services, namely `Nested`, mutates the request URI so we must
+        // restore it to its original state before calling `second`
+        original_uri: Option<http::Uri>,
     }
 }
 
@@ -100,7 +105,7 @@ where
                 StateProj::FirstFuture { f } => {
                     let mut response = ready!(f.poll(cx)?);
 
-                    let req = if let Some(ext) = response
+                    let mut req = if let Some(ext) = response
                         .extensions_mut()
                         .remove::<FromEmptyRouter<ReqBody>>()
                     {
@@ -108,6 +113,8 @@ where
                     } else {
                         return Poll::Ready(Ok(response));
                     };
+
+                    *req.uri_mut() = this.original_uri.take().unwrap();
 
                     let second = this.second.take().expect("future polled after completion");